Don't segment using clips with different positioning nodes
When segmenting brushes, don't consider clips if they have a different
positioning node. The issue here is that since they have a different
positioning node, the segmentation results may change during scrolling.
This means that we would need to invalidate the cache result and
re-segment. We can do that in the future, but in the meantime this fixes
the rendering issues the bug causes.

Unfortunately, this bug is triggered by rendering, scrolling, and then
re-rendering, so it is tricky to make a test for it.

Original bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1446626